Modularize BodyEditorLoader
