+++ core/src/com/fourbit/pc_invader/entities/boss/Boss.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.physics.box2d.BodyDef;
 import com.badlogic.gdx.physics.box2d.FixtureDef;
 import com.badlogic.gdx.physics.box2d.World;
 import com.badlogic.gdx.utils.Array;
-import com.fourbit.pc_invader.BodyEditorLoader;
+import com.fourbit.pc_invader.utils.BodyEditorLoader;
 import com.fourbit.pc_invader.entities.Entity;
